+++ packages/smarthr-ui/src/components/Combobox/useListbox.tsx
@@ -26,7 +26,6 @@ import { Text } from '../Text'
 import { VisuallyHiddenText } from '../VisuallyHiddenText'
 
 import { ItemButton } from './ItemButton'
-import { useActiveOption } from './useActiveOption'
 
 import type { ComboboxItem, ComboboxOption } from './types'
 
@@ -96,7 +95,47 @@ export const useListbox = <T,>({
   noResultText,
 }: Props<T>) => {
   const [navigationType, setNavigationType] = useState<'pointer' | 'key'>('pointer')
-  const { activeOption, setActiveOption, moveActiveOptionIndex } = useActiveOption({ options })
+  const [activeOption, setActiveOption] = useState<ComboboxOption<T> | null>(null)
+
+  useEffect(() => {
+    // props の変更によって activeOption の状態が変わりうるので、実態を反映する
+    setActiveOption((current) => {
+      if (current === null) {
+        return null
+      }
+
+      return options.find((option) => current.id === option.id) ?? null
+    })
+  }, [options])
+
+  const moveActiveOptionIndex = useCallback(
+    (currentActive: ComboboxOption<T> | null, delta: -1 | 1) => {
+      if (options.every((option) => option.item.disabled)) {
+        return
+      }
+
+      const currentActiveIndex =
+        currentActive === null ? -1 : options.findIndex((option) => option.id === currentActive.id)
+      let nextIndex = 0
+
+      if (currentActiveIndex !== -1) {
+        nextIndex = (currentActiveIndex + delta + options.length) % options.length
+      } else if (delta !== 1) {
+        nextIndex = options.length - 1
+      }
+
+      const nextActive = options[nextIndex]
+
+      if (nextActive) {
+        if (nextActive.item.disabled) {
+          moveActiveOptionIndex(nextActive, delta)
+        } else {
+          setActiveOption(nextActive)
+        }
+      }
+    },
+    [options],
+  )
 
   useEffect(() => {
     // 閉じたときに activeOption を初期化
